Fix gcc 11 / libstdc++ CI errors
Two distinct build errors surfaced on the Linux gcc-11 CI run:

1. STRONG_TYPEDEF_EXPLICIT value-ctors were marked constexpr, but
   std::string's copy/move constructors are NOT constexpr in libstdc++
   (only in libc++). Apple builds happened to compile because libc++
   relaxes this. Drop constexpr from both value ctors.

2. std::to_string(uint8_t) / std::to_string(uint16_t) trigger
   -Werror=sign-promo on gcc because the integer-promotion candidates
   include both int and unsigned int. Cast through unsigned to pick
   the unambiguous overload.

+++ components/catalog/system_table_schemas.cpp
@@ -352,7 +352,8 @@ namespace components::catalog {
 
         if (t.type() == LT::DECIMAL) {
             const auto* ext = static_cast<const types::decimal_logical_type_extension*>(t.extension());
-            return "numeric(" + std::to_string(ext->width()) + "," + std::to_string(ext->scale()) + ")";
+            return "numeric(" + std::to_string(static_cast<unsigned>(ext->width())) + "," +
+                   std::to_string(static_cast<unsigned>(ext->scale())) + ")";
         }
         if (t.type() == LT::UNKNOWN) {
             return "UNKNOWN(" + t.type_name() + ")";
@@ -709,9 +710,9 @@ namespace components::catalog {
                 out += v.value<bool>() ? '1' : '0';
                 break;
             case LT::TINYINT:   out += std::to_string(v.value<int8_t>());   break;
-            case LT::UTINYINT:  out += std::to_string(v.value<uint8_t>());  break;
+            case LT::UTINYINT:  out += std::to_string(static_cast<unsigned>(v.value<uint8_t>()));  break;
             case LT::SMALLINT:  out += std::to_string(v.value<int16_t>());  break;
-            case LT::USMALLINT: out += std::to_string(v.value<uint16_t>()); break;
+            case LT::USMALLINT: out += std::to_string(static_cast<unsigned>(v.value<uint16_t>())); break;
             case LT::INTEGER:   out += std::to_string(v.value<int32_t>());  break;
             case LT::UINTEGER:  out += std::to_string(v.value<uint32_t>()); break;
             case LT::BIGINT:    out += std::to_string(v.value<int64_t>());  break;
